backend/requirements.txt
autopep8
django-cors-headers==3.10.0
Django==3.2.9
djangorestframework==3.12.4
python-dateutil
python-dotenv